Refactor casts, statics and minor cleanups
Apply a set of small refactors and cleanup across multiple scripts:
- Ensure floating-point division by casting to float in SessionMatchLobbyGameAdapter to produce correct progress fractions.
- Make several helper methods/static checks static where safe (PostMatchManager.ResolveWorldCamera, WeaponMount.TryResolveMainSceneCamera, CustomNetworkManager.ApprovalCheck) and rename a private flag to camelCase in SessionManager.
- Simplify control flow and pattern matching (SessionSceneFlowGameAdapter ternary, ChatUIManager pattern match) and remove unused usings.
- Convert VoiceManager.Start from async void to launching an async Task to avoid async void pitfalls.
- Simplify and harden ResolveMaxPlayersForMode in SessionMatchmaker (use concise ternary, early return and clamp invalid results to 1 with a debug warning).
These changes improve correctness, readability, and safety without altering behavior intent.

+++ Assets/Scripts/Game/Adapters/SessionMatchLobbyGameAdapter.cs
@@ -37,7 +37,7 @@ namespace Game.Adapters {
                 if(timer.IsWaitingForPlayers || timer.IsPreMatch) return (true, "PreMatch");
                 var duration = matchSettings != null ? matchSettings.GetMatchDurationSeconds() : 0;
                 if(duration > 0 && timer.TimeRemainingSeconds >= 0) {
-                    var remainingFraction = timer.TimeRemainingSeconds / Mathf.Max(1, duration);
+                    var remainingFraction = timer.TimeRemainingSeconds / (float)Mathf.Max(1, duration);
                     var minRemaining = ResolveBackfillTimeRemainingThreshold(mode);
                     if(remainingFraction <= minRemaining) return (false, $"LateTime:{remainingFraction:0.00}");
                 }
@@ -90,7 +90,7 @@ namespace Game.Adapters {
             var scoreToWin = MatchSettingsManager.Instance != null ? MatchSettingsManager.Instance.GetScoreToWin() : 0;
             if(scoreToWin <= 0) return 0f;
             var leadingScore = Mathf.Max(teamAScore, teamBScore);
-            return leadingScore / Mathf.Max(1, scoreToWin);
+            return leadingScore / (float)Mathf.Max(1, scoreToWin);
         }
 
         private static float ResolveLeadingTeamObjectiveOrKillProgress(bool teamMode) {
@@ -103,7 +103,7 @@ namespace Game.Adapters {
                     if(player == null || player.NetworkObject == null || !player.NetworkObject.IsSpawned) continue;
                     leadingKills = Mathf.Max(leadingKills, player.Kills.Value);
                 }
-                return leadingKills / Mathf.Max(1, scoreToWin);
+                return leadingKills / (float)Mathf.Max(1, scoreToWin);
             }
 
             var teamAKills = 0;
@@ -119,7 +119,7 @@ namespace Game.Adapters {
                     default: throw new ArgumentOutOfRangeException();
                 }
             }
-            return Mathf.Max(teamAKills, teamBKills) / Mathf.Max(1, scoreToWin);
+            return Mathf.Max(teamAKills, teamBKills) / (float)Mathf.Max(1, scoreToWin);
         }
     }
 }
